[tests] Bug 77 renamed issues test bundle and fixed test failures
