--- plugins/org.locationtech.udig.issues.test/fragment.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.2"?>
-<fragment>
-   <extension
-         point="org.locationtech.udig.issues.issuesViewConfiguration">
-      <contentProvider
-            class="org.locationtech.udig.issues.test.TestContentProvider$Provider1"
-            id="TestContentProvider"/>
-      <expansionProvider
-            class="org.locationtech.udig.issues.test.TestExpansionProvider$Provider1"
-            id="TestContentProvider"/>
-      <sorter
-            class="org.locationtech.udig.issues.test.TestSorter$Sorter1"
-            id="TestSorter"/>
-      <labelProvider
-            class="org.locationtech.udig.issues.test.TestLabelProvider$Provider1"
-            id="TestLabelProvider"/>
-   </extension>
-   <extension
-         point="org.locationtech.udig.issues.issueFixer">
-      <fixer
-            class="org.locationtech.udig.issues.test.NonExistentClass"
-            id="TestIssueFixer1">
-         <requiredKey key="missingKey"/>
-      </fixer>
-      <fixer
-            class="org.locationtech.udig.issues.test.NonExistentClass"
-            id="TestIssueFixer2"
-            targetClass="org.locationtech.udig.issues.test.DummyFixableIssue"/>
-      <fixer
-            class="org.locationtech.udig.issues.test.DummyAutoIssueFixer"
-            id="TestIssueFixer3"
-            targetClass="org.locationtech.udig.issues.FixableIssue">
-         <requiredKey key="fixable"/>
-      </fixer>
-   </extension>
-
-</fragment>
