--- plugins/org.locationtech.udig.issues.test/pom.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-    xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <artifactId>plugins</artifactId>
-        <groupId>org.locationtech.udig</groupId>
-        <version>2.0.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>org.locationtech.udig.issues.test</artifactId>
-    <packaging>eclipse-test-plugin</packaging>
-
-    <build> 
-        <plugins>
-            <plugin>
-                <groupId>org.eclipse.tycho</groupId>
-                <artifactId>tycho-surefire-plugin</artifactId>
-                <configuration>
-                    <testRuntime>p2Installed</testRuntime>
-                </configuration>
-            </plugin>
-
-            <plugin>
-                <groupId>org.eclipse.tycho</groupId>
-                <artifactId>target-platform-configuration</artifactId>
-                <configuration>
-                    <dependency-resolution>
-                        <extraRequirements>
-                            <requirement>
-                                <type>p2-installable-unit</type>
-                                <id>org.locationtech.udig-product</id>
-                                <versionRange>0.0.0</versionRange>
-                            </requirement>
-                          <requirement>
-                             <type>eclipse-feature</type>
-                             <id>org.eclipse.e4.rcp</id>
-                             <versionRange>0.0.0</versionRange>
-                          </requirement>
-                        </extraRequirements>
-                    </dependency-resolution>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
