+++ plugins/org.locationtech.udig.issues.tests/.classpath
