+++ plugins/org.locationtech.udig.issues.tests/META-INF/MANIFEST.MF
@@ -1,0 +1,15 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Issues Test Plug-in
+Bundle-SymbolicName: org.locationtech.udig.issues.tests;singleton:=true
+Bundle-Version: 2.0.0.qualifier
+Bundle-Vendor: udig.refractions.net
+Require-Bundle: org.eclipse.ui,
+ org.eclipse.core.runtime,
+ org.locationtech.udig.core.testsupport,
+ org.locationtech.udig.project.tests.ui,
+ org.locationtech.udig.catalog.testsupport,
+ org.locationtech.udig.project.testsupport
+Bundle-ActivationPolicy: lazy
+Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Fragment-Host: org.locationtech.udig.issues
