+++ plugins/org.locationtech.udig.issues.tests/build.properties
