+++ plugins/org.locationtech.udig.issues.tests/src/org/locationtech/udig/issues/DummyListener.java
@@ -1,0 +1,16 @@
+/**
+ * 
+ */
+package org.locationtech.udig.issues;
+
+import org.locationtech.udig.issues.listeners.IIssuesListListener;
+import org.locationtech.udig.issues.listeners.IssuesListEvent;
+
+class DummyListener implements IIssuesListListener{
+    int changes=0;
+    int timesCalled;
+    public void notifyChange( IssuesListEvent event ) {
+        changes+=event.getChanged().size();
+        timesCalled++;    
+    }
+}
